@@ -8,28 +8,45 @@
   },
   "main": "js/app.js",
   "dependencies": {
-    "babel-core": "^5.8.25",
-    "babel-loader": "^5.3.2",
-    "expose-loader": "^0.7.0",
-    "history": "^1.13.1",
+    "babel": "^6.3.26",
+    "babel-cli": "^6.3.17",
+    "babel-core": "^6.3.26",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-syntax-decorators": "^6.3.13",
+    "babel-polyfill": "^6.3.14",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "es5-shim": "^4.4.1",
+    "expose-loader": "^0.7.1",
+    "history": "^1.17.0",
     "node-libs-browser": "^0.5.3",
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0",
-    "react-redux": "^3.1.0",
-    "react-router": "^1.0.0",
-    "redux": "^3.0.2",
-    "redux-thunk": "^1.0.0",
-    "webpack": "^1.12.2"
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3",
+    "react-on-rails": "../../..",
+    "react-redux": "^4.0.5",
+    "react-router": "^1.0.3",
+    "redux": "^3.0.5",
+    "redux-thunk": "^1.0.2",
+    "webpack": "^1.12.8"
   },
   "devDependencies": {
-    "babel-eslint": "^4.1.3",
-    "eslint": "^1.6.0",
-    "eslint-config-airbnb": "0.1.0",
-    "eslint-plugin-react": "^3.5.1",
-    "jscs": "^2.3.0"
+    "babel-eslint": "^5.0.0-beta6",
+    "eslint": "^1.10.3",
+    "eslint-config-airbnb": "2.1.1",
+    "eslint-plugin-react": "^3.13.1",
+    "jscs": "^2.8.0"
+  },
+  "browser": {
+    "fs": false
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "npm run eslint --silent && npm run jscs --silent",
+    "eslint": "eslint --ext .js,.jsx .",
+    "jscs": "jscs --verbose .",
+    "jscs:fix": "jscs --verbose -x . --silent",
+    "build:test":  "webpack --config webpack.client.js && webpack --config webpack.server.js",
+    "test": "npm run build:test && rspec",
     "build:dev:client": "webpack -w --config webpack.client.js",
     "build:dev:server": "webpack -w --config webpack.server.js"
   },
