@@ -1,9 +1,9 @@
 import React from 'react';
-import {Link} from 'react-router';
+import { Link } from 'react-router';
 
 export default class RouterLayout extends React.Component {
   static propTypes = {
-    children: React.PropTypes.object
+    children: React.PropTypes.object,
   };
 
   render() {
