@@ -49,7 +49,7 @@ var server = new WebpackDevServer(webpack(config), {
 
 var initialName = 'Stranger';
 
-server.app.use('/', function(req, res) {
+server.app.use('/', function routePath(req, res) {
   var locals = {
     props: JSON.stringify(initialName),
   };
@@ -58,7 +58,7 @@ server.app.use('/', function(req, res) {
   res.send(html);
 });
 
-server.listen(4000, 'localhost', function(err) {
+server.listen(4000, 'localhost', function onListen(err) {
   if (err) console.log(err);
   console.log('Listening at localhost:4000...');
 });
