@@ -14,18 +14,30 @@ config.output = {
 };
 
 // You can add entry points specific to rails here
+// The es5-shim/sham is for capybara testing
 config.entry.vendor.unshift(
   'es5-shim/es5-shim',
   'es5-shim/es5-sham'
 );
 
 // jquery-ujs MUST GO AFTER jquery, so must use 'push'
-config.entry.app.push('jquery-ujs');
+config.entry.vendor.push('jquery-ujs');
 
 // See webpack.common.config for adding modules common to both the webpack dev server and rails
 config.module.loaders.push(
-  {test: /\.jsx?$/, loader: 'babel-loader', exclude: /node_modules/},
-  {test: require.resolve('react'), loader: 'imports?shim=es5-shim/es5-shim&sham=es5-shim/es5-sham'}
+  {
+    test: /\.jsx?$/,
+    loader: 'babel-loader',
+    exclude: /node_modules/,
+  },
+  {
+    test: require.resolve('react'),
+    loader: 'imports?shim=es5-shim/es5-shim&sham=es5-shim/es5-sham',
+  },
+  {
+    test: require.resolve('jquery-ujs'),
+    loader: 'imports?jQuery=jquery',
+  }
 );
 
 module.exports = config;
