@@ -10,13 +10,16 @@ module.exports = {
 
   // the project dir
   context: __dirname,
-  entry: ['./app/bundles/HelloWorld/startup/serverGlobals', 'react', 'react-dom/server'],
+  entry: [
+    'babel-polyfill',
+    './app/bundles/HelloWorld/startup/serverRegistration',
+  ],
   output: {
     filename: 'server-bundle.js',
     path: '../app/assets/javascripts/generated',
   },
   resolve: {
-    extensions: ['', '.webpack.js', '.web.js', '.js', '.jsx', 'config.js'],
+    extensions: ['', '.js', '.jsx'],
     alias: {
       lib: path.join(process.cwd(), 'app', 'lib'),
     },
@@ -24,17 +27,13 @@ module.exports = {
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
-        NODE_ENV: JSON.stringify('production'),
+        NODE_ENV: JSON.stringify(nodeEnv),
       },
     }),
   ],
   module: {
     loaders: [
-      {test: /\.jsx?$/, loader: 'babel-loader', exclude: /node_modules/},
-
-      // React is necessary for the client rendering:
-      {test: require.resolve('react'), loader: 'expose?React'},
-      {test: require.resolve('react-dom/server'), loader: 'expose?ReactDOMServer'},
+      { test: /\.jsx?$/, loader: 'babel-loader', exclude: /node_modules/ },
     ],
   },
 };
