@@ -4,29 +4,31 @@ import _ from 'lodash';
 
 // Simple example of a React "smart" component
 export default class HelloWorld extends React.Component {
+  static propTypes = {
+    name: PropTypes.string.isRequired, // this is passed from the Rails view
+  };
+
   constructor(props, context) {
     super(props, context);
+
+    // How to set initial state in ES6 class syntax
+    // https://facebook.github.io/react/docs/reusable-components.html#es6-classes
+    this.state = { name: this.props.name };
 
     // Uses lodash to bind all methods to the context of the object instance, otherwise
     // the methods defined here would not refer to the component's class, not the component
     // instance itself.
-    _.bindAll(this, '_updateName');
+    _.bindAll(this, 'updateName');
   }
 
-  static propTypes = {
-    name: PropTypes.string.isRequired, // this is passed from the Rails view
-  }
-
-  state = {name: this.props.name} // how to set initial state in es2015 class syntax
-
-  _updateName(name) {
-    this.setState({name: name});
+  updateName(name) {
+    this.setState({ name });
   }
 
   render() {
     return (
       <div>
-        <HelloWorldWidget name={this.state.name} _updateName={this._updateName} />
+        <HelloWorldWidget name={this.state.name} updateName={this.updateName} />
       </div>
     );
   }
