@@ -1,11 +1,8 @@
 import React from 'react';
 import HelloWorld from '../containers/HelloWorld';
 
-const HelloWorldAppServer = props => {
-  const reactComponent = (
+export default (props) => {
+  return (
     <HelloWorld {...props} />
   );
-  return reactComponent;
 };
-
-export default HelloWorldAppServer;
