@@ -20,7 +20,7 @@ export default props => {
   // Redux expects to initialize the store using an Object, not an Immutable.Map
   const initialState = {
     $$helloWorldStore: $$helloWorldState.merge({
-      name: name,
+      name,
     }),
   };
 
