@@ -8,7 +8,7 @@ import HelloWorld from '../containers/HelloWorld';
 // This is how you get props from the Rails view into the redux store.
 // This code here binds your smart component to the redux store.
 // This is how the server redux gets hydrated with data.
-const HelloWorldAppServer = props => {
+export default (props) => {
   const store = createStore(props);
   const reactComponent = (
     <Provider store={store}>
@@ -17,5 +17,3 @@ const HelloWorldAppServer = props => {
   );
   return reactComponent;
 };
-
-export default HelloWorldAppServer;
