@@ -1,3 +1,3 @@
 module ReactOnRails
-  VERSION = "1.2.2"
+  VERSION = "2.0.0.rc.3"
 end
