@@ -13,7 +13,7 @@ const HelloWorldES5 = React.createClass({
 
   _handleChange() {
     const name = ReactCompat.reactFindDOMNode()(this.refs.name).value;
-    this.setState({name});
+    this.setState({ name });
   },
 
   render() {
