@@ -18,7 +18,7 @@ import HelloWorldContainer from '../components/HelloWorldContainer';
  *  React will see that the state is the same and not do anything.
  *
  */
-const ReduxApp = props => {
+export default props => {
   const combinedReducer = combineReducers(reducers);
   const store = applyMiddleware(middleware)(createStore)(combinedReducer, props);
 
@@ -39,12 +39,3 @@ const ReduxApp = props => {
 
   return reactComponent;
 };
-
-/*
- * If you wish to create a React component via a function, rather than simply props,
- * then you need to set the property "generator" on that function to true.
- * When that is done, the function is invoked with a single parameter of "props",
- * and that function should return a react element.
- */
-
-export default ReduxApp;
