@@ -15,5 +15,6 @@ Rails.application.routes.draw do
   get "server_side_hello_world_with_options" => "pages#server_side_hello_world_with_options"
   get "server_side_redux_app_cached" => "pages#server_side_redux_app_cached"
   get "render_js" => "pages#render_js"
-  get "react_router(/*all)" => "react_router#index"
+  get "react_router(/*all)" => "react_router#index", as: :react_router
+  get "pure_component" => "pages#pure_component"
 end
