@@ -1,7 +1,6 @@
 shared_examples "no_redux_generator:base" do
   it "copies non-redux base files" do
-    %w(client/app/bundles/HelloWorld/components/HelloWorldWidget.jsx
-       client/app/bundles/HelloWorld/containers/HelloWorld.jsx).each { |file| assert_file(file) }
+    assert_file("client/app/bundles/HelloWorld/containers/HelloWorld.jsx")
   end
 
   it "does not place react folders in root" do
@@ -18,9 +17,8 @@ shared_examples "no_redux_generator:no_server_rendering" do
   end
 
   it "templates the client-side-rendering version of HelloWorldApp" do
-    assert_file("client/app/bundles/HelloWorld/startup/HelloWorldApp.jsx") do |contents|
-      assert_match("HelloWorldApp", contents)
-      refute_match("HelloWorldAppClient", contents)
+    assert_file("client/app/bundles/HelloWorld/startup/HelloWorldAppClient.jsx") do |contents|
+      assert_match("HelloWorld", contents)
     end
   end
 end
@@ -33,7 +31,7 @@ shared_examples "no_redux_generator:server_rendering" do
 
   it "templates the server-side-rendering version of HelloWorldApp" do
     assert_file("client/app/bundles/HelloWorld/startup/HelloWorldAppClient.jsx") do |contents|
-      assert_match("HelloWorldAppClient", contents)
+      assert_match("HelloWorld", contents)
     end
   end
 end
